feat(clinical-details) loader shows when fetching and saving form data
